Rename domains header
